New map and map merging
This commit update issue #2 #3 #4 #5.

SLAM is better but the problem from issue #8 remains.

The robots are not bumping into each other thanks to gmapping correct positioning.

The map-merged is now working and the AMCL is based on it.

The position of the map has been offset to match the real map. A manual change has been been made into the m-explore, branch publish-transform, merging_pipeline.cpp, line 222 and 223.

The new world has now the same resolution as the map. 0 is now the center of the of the map and not the bottom left corner.