Push for change of resolution to 0.05
This is a branch to test out solution of issue #10.

This is a test push for changes made in order to respect the 0.05 resolution of gmapping. The map seems too big though. A new map has to be done to counter the effect if the 0.01 resolution is to be kept.