Changes when loading the map and adding parameters for gmapping
When loading the empty map the file didn't exist.

Two more parameters into gmapping. A vain attempt to change to position of the map. Gmapping package has been changed as well. Line 712 and 713 of the slam_gmapping.cpp file of the gmapping package added the parameter x/ycenter fetch (+ otehr modification on the top of the file and inside the header).