movebaseHandler updated with a scan of proximity cells of the goal
The cells next to the pixel of the goal are scanned to find a free cell near the goal.
The idea is to verify if the goal is on a free cell.
If not, a scan of the neighbor will occur (like a snail shell).
This scan will occur with 10 incrementing scan.