movebaseHandler map debugging tools
This commit updates issue #11.

In order to help debugging, publishing of the maps and point have been set up.

Two different strategies have been put in place:
- analysing the square (passing over already analysed  pixels)
- analysing in snake (only do rings and increments)

To help visualize the map, map saver must be used on robot_0/new_map

Problem with the exact number of pixels, the position of start of the analysis is wrong. It seems shifted to the right.