Add script for patrolling
This commit updates issue #13.

A goal is taken from a list of available goal.
The goal is removed from the list.
This script is sending the value to the simple_goal topic of the robot.
Once a goal has been visited (verified thanks to amcl_pose), the goal is added at the end of the free goal list.
So on for each robot. The list is shared among all robots.